(BKR-623) Specialize Beaker::TestSuite
This creates a `lib/beaker/runnner` path, a `Beaker::Runner` module namespace,
and moves the `Beaker::TestSuite` there, leaving a wrapper subclass behind.

This allows us to treat the current Beaker test runner as a special case, making
space to add support for other test runners. Everything should continue working
as before, allowing us to move on to next steps.
